rename headers for nos_ prefix
