+++ core/include/nos_core.h
@@ -1,13 +1,13 @@
 #ifndef __NOS_CORE_H__
 #define __NOS_CORE_H__
 
-#include <file_intf.h>
-#include <raw_socket.h>
-#include <time_intf.h>
-#include <evt_mgr_intf.h>
+#include <nos_file_intf.h>
+#include <nos_raw_socket.h>
+#include <nos_time_intf.h>
+#include <nos_evt_mgr_intf.h>
 #include <nos_logging.h>
 #include <nos_log_factory.h>
-#include <udp_socket_intf.h>
+#include <nos_udp_socket_intf.h>
 #include <nos_log_msg_intf.h>
 
 #endif
