+++ core/include/nos_evt_mgr_intf.h
