+++ core/include/nos_file_intf.h
