+++ core/include/nos_tcp_socket_intf.h
@@ -1,0 +1,25 @@
+#ifndef __NOS_TCP_SOCKET_INTF_H__
+#define __NOS_TCP_SOCKET_INTF_H__
+
+namespace nos::core
+{
+
+class tcp_server {
+
+};
+
+class tcp_client {
+
+};
+
+class tcp_unix_server {
+
+};
+
+class tcp_unix_client {
+
+};
+
+}
+
+#endif
