+++ core/src/udp_socket_intf.cc
@@ -12,7 +12,7 @@
 #include <sys/un.h>
 #include <arpa/inet.h>
 #include <netinet/in.h>
-#include <udp_socket_intf.h>
+#include <nos_udp_socket_intf.h>
 
 namespace nos::core
 {
